requirements.txt
fastf1>=3.4.4
requests>=2.32.0
PyYAML>=6.0.1
pydantic>=2.7.0
pulp>=2.8.0
markdown>=3.6
python-dateutil>=2.9.0
pytest>=8.2.0
